requirements.txt
google-genai
streamlit
Pillow
python-dotenv